some changes - move Info and Student to packages
