+++ src/main/java/ru/hogwarts/school/controller/AvatarController.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 import ru.hogwarts.school.model.Avatar;
 import ru.hogwarts.school.service.AvatarService;
-import ru.hogwarts.school.service.StudentService;
+import ru.hogwarts.school.service.student.StudentService;
 
 import java.io.IOException;
 import java.util.Collection;
