+++ src/main/java/ru/hogwarts/school/controller/InfoController.java
@@ -3,7 +3,7 @@ package ru.hogwarts.school.controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import ru.hogwarts.school.configuration.InfoService;
+import ru.hogwarts.school.service.info.InfoService;
 
 @RestController
 @RequestMapping("/port")
