+++ src/main/java/ru/hogwarts/school/controller/StudentController.java
@@ -6,8 +6,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import ru.hogwarts.school.model.Faculty;
 import ru.hogwarts.school.model.Student;
-import ru.hogwarts.school.service.StudentParallelPrinter;
-import ru.hogwarts.school.service.StudentService;
+import ru.hogwarts.school.service.student.StudentParallelPrinter;
+import ru.hogwarts.school.service.student.StudentService;
 
 import java.util.Collection;
 import java.util.List;
